Source Code Library Update
Updated at 22:55.

Well, it's yet another end of the weekend. Aha, how time flies. Did
update some of the APPs on my cellphone today. Anyhow, time to update my
cellphone then, so long that I have got the money I needed.

Let's make a summary them.
-------
Did program some MATLAB codes, according to the text books. But that
shall never be enough. And played for some more time that EXPECTED over
the May 1st holidays.

Well, have to work harder to make up for the lost time & work loads,
that's just another Old Saying......But it's TRUE, anyhow.

-------
Try to get in touch with those teachers over the sea and find more about
that fucking country as described and appraised......

Looking forward to a Ph.D. life then, for Geeks or Nerds like me, it may
not be the best choice, but it shall never be the worst one, isn;t it?

Keep working, keep living.
Life as a bitch. Fuck it till explode.