Source Code Library Update
Updated at 22:55.

Well, yes I have to say when I update those files and write a comment
like this it's JUST another END OF THE DAY.
-------
There's a saying that, "After all, tomorrow is another day."

Like the guys who have chosen to be a postgraduate student in NWPU, you
have to work harder as a Ph.D. student abroad and, if anything, with a
Assistant funding...... Aha, without the money go hell the Research &
Development.

We China, is just a developing country, which means that it's
developing, and we have to work as (not as if) cheap labours to develop
it. Or? Or rather it will stay still and be a nation like Japna.

Will you want to see your so-called "Mother Country" be like Fucking
Japan? No! Out of it!!!

So? So take some more exercises and be healthy enough, listen well to be
more agust. And be back tomorrow to be a more adept programmer,
flowershly named,
Software Engineer.

Qi Lai, Bu Yuan Zuo Nu Li De Ren Men.

-------
Crazy words by Abraham Xiao
@ InfoSec Lab,
School of Automation,
Northwestern Polytechnic University.