Source Code Library Update
Updated at 17:40.
-------
Well, when moments come to write descriptions to one commit, it's either
close to the lunch time, or dinner time, or, you name it, sleeping time.
How time flies when you stay in front of a computer, or a book. Either
programming or reading is fun, both of them take a lot of time to
accomplish.
-------
When I commit this and synchronize it to the server, my own account,
it's almost the end of the bright day. You know it could give you so
much strength and/or power looking at the funny & crappy post-graduates,
especially those guys/pals who are Non-NPUers. Judging from their
background, their motivation to learn, their REAL-TIME actions in the
Lab, in the classrooms, on bed...Aha, those bitches.
-------
They and some of your peers could act as a mirror. They are always
reminding me when to work, when to play and when to sleep. They name
this peer pressure, I name it peer pleasure.
-------
Abraham! Do something, make a difference.
Stop fucking those words out of your mouth, go and take some trash as my
dinner, attend a lecture of Collean Lavery, a GREAT female foreign
teacher you can not always find in NWPU. Go and have some discussions,
then be back, be a programmer for approximately two hours.
-------
After all and known to all, LIFE LIES IN PHYSICAL EXERCISE.
Take the fucking jogging the night as usual, read something, you poor
Non-Hacker. Sleep. Tomorrow is another day.
-------
By Abraham Xiao @ InfoSec Lab,
School of Automation,
Northwestern Polytechnic University.
Xi'an, Shaanxi Province,
People's Republic of Capitalism.